refactor: improve npkill file
