+++ src/core/interfaces/index.ts
@@ -1,0 +1,4 @@
+export * from './file-service.interface.js';
+export * from './folder.interface.js';
+export * from './services.interface.js';
+export * from './search-status.model.js';
