+++ src/core/interfaces/services.interface.ts
@@ -1,0 +1,14 @@
+import { FileService, FileWorkerService } from '@core/services/files/index.js';
+import { LoggerService } from '@core/services/logger.service.js';
+import { StreamService } from '@core/services/stream.service.js';
+import { ResultsService } from 'src/cli/services/index.js';
+import { SearchStatus } from './search-status.model.js';
+
+export interface Services {
+  logger: LoggerService;
+  searchStatus: SearchStatus;
+  fileService: FileService;
+  fileWorkerService: FileWorkerService;
+  streamService: StreamService;
+  resultsService: ResultsService;
+}
